dpdk: Make hugepage size a changeable global parameter
Currently the DPDK testcase is hardcoded to use 1GiB hugepages, and that
assumption is scattered across multiple places in the code.  Make this more
flexible by making the hugepage size a global variable, initialized once,
from which we base everything else.

This is slightly complicated because different systems use different
representations of the the hugepagesize.

Signed-off-by: David Gibson <dgibson@redhat.com>