dpdk: Compute global hugepage requirements from per-pod requirements
As well as assigning hugepage memory to individual pods, the DPDK test case
uses a performance profile to manage the pool of hugepages globally for the
k8s node.  Currently that's hardcoded to allocate 10 pages, which only
makes sense for the current particular page size and per-pod allocation
values.

Since those are now adjustable (as set-once global variables), compute the
host-wide requirements from the page size and per-pod requirements.

Signed-off-by: David Gibson <dgibson@redhat.com>